+++ R/CDISCFilteredData.R
@@ -54,10 +54,9 @@
 #'
 #' isolate({
 #'   datasets$datanames()
-#'   datasets$get_data_info("ADSL", filtered = FALSE)
 #'
-#'   # filters dataset to obtain information
-#'   datasets$get_data_info("ADSL", filtered = TRUE)
+#'   # number observations and subjects of filtered/non-filtered dataset
+#'   datasets$get_filter_overview("ADSL")
 #'
 #'   print(datasets$get_call("ADSL"))
 #'   print(datasets$get_call("ADTTE"))
@@ -141,27 +140,6 @@ CDISCFilteredData <- R6::R6Class( # nolint
         self$get_filtered_datasets(dataname)$get_call()
       }
     },
-
-    #' @description
-    #' Get info about dataname, i.e. number of rows, subjects.
-    #'
-    #' @param dataname (`character`) name of the dataset
-    #' @param filtered (`logical`) whether to obtain this info for the
-    #'   filtered dataset
-    #' @return a named vector
-    get_data_info = function(dataname, filtered) {
-      private$check_data_varname_exists(dataname)
-      stopifnot(is_logical_single(filtered))
-
-      nrows <- self$get_filtered_datasets(dataname)$get_data_info(filtered = filtered)
-      nsubjects <- self$get_filtered_datasets(dataname)$get_subjects_info(filtered = filtered)
-
-      list(
-        Obs = nrows,
-        Subjects = nsubjects
-      )
-    },
-
 
     #' @description
     #' Get names of datasets available for filtering