+++ R/FilteredData.R
@@ -47,10 +47,10 @@
 #'
 #' isolate({
 #'   datasets$datanames()
-#'   datasets$get_data_info("iris", filtered = FALSE)
+#'   datasets$get_filter_overview("iris")
 #'
 #'   # filters dataset to obtain information
-#'   datasets$get_data_info("mtcars", filtered = TRUE)
+#'   datasets$get_filter_overview("mtcars")
 #'
 #'   print(datasets$get_call("iris"))
 #'   print(datasets$get_call("mtcars"))
@@ -261,64 +261,42 @@ FilteredData <- R6::R6Class( # nolint
 
       return(res)
     },
-
-
+    
     #' @description
     #' Get filter overview table in form of X (filtered) / Y (non-filtered)
     #'
     #' This is intended to be presented in the application.
-    #' The content for each of the data names is defined in `get_data_info` method.
+    #' The content for each of the data names is defined in `get_filter_overview_info` method.
     #'
     #' @param datanames (`character` vector) names of the dataset
-    #' @return (`data.frame`)
-    get_filter_overview_tbl = function(datanames) {
+    #'
+    #' @return (`matrix`) matrix of observations and subjects of all datasets
+    get_filter_overview = function(datanames) {
       if (identical(datanames, "all")) {
         datanames <- self$datanames()
       }
-
       check_in_subset(datanames, self$datanames(), "Some datasets are not available: ")
 
-      data_info_filtered <- sapply(
+      rows <- sapply(
         datanames,
         function(dataname) {
-          self$get_data_info(dataname, filtered = TRUE)
-        },
-        USE.NAMES = TRUE,
-        simplify = FALSE
-      )
-      data_info_nfiltered <- sapply(
-        datanames,
-        function(dataname) {
-          self$get_data_info(dataname, filtered = FALSE)
-        },
-        USE.NAMES = TRUE,
-        simplify = FALSE
-      )
-
-      res_list <- Map(
-        x = data_info_filtered,
-        y = data_info_nfiltered,
-        function(x, y) {
-          setNames(paste0(x, "/", y), names(x))
-          sapply(
-            names(x),
-            function(xname) {
-              paste0(x[[xname]], "/", y[[xname]])
-            },
-            USE.NAMES = TRUE
-          )
+          self$get_filtered_datasets(dataname)$get_filter_overview_info()
         }
       )
 
-      res_df <- as.data.frame(do.call(rbind, res_list))
-      res_df[, "Dataset"] <- rownames(res_df)
-      rownames(res_df) <- NULL
-      res_df <- res_df[, c("Dataset", setdiff(names(res_df), "Dataset"))]
-
-      return(res_df)
+      do.call(rbind, rows)
     },
 
-
+    #' @description
+    #' Get join keys between two datasets.
+    #' @param dataset_1 (`character`) one dataset name
+    #' @param dataset_2 (`character`) other dataset name
+    #' @return (`named character`) vector with column names
+    get_join_keys = function(dataset_1, dataset_2) {
+      if (is.null(private$join_keys))
+        return(character(0))
+      private$join_keys$get(dataset_1, dataset_2)
+    },
 
     #' Get keys for the dataset
     #' @param dataname (`character`) name of the dataset
@@ -697,7 +675,7 @@ FilteredData <- R6::R6Class( # nolint
       ns <- NS(id)
 
       div(
-        style = "overflow: overlay",
+        class = "teal_active_summary_filter_panel",
         tableOutput(ns("table"))
       )
     },
@@ -718,7 +696,7 @@ FilteredData <- R6::R6Class( # nolint
         is.function(active_datanames) || is.reactive(active_datanames)
       )
 
-      output$table <- renderTable({
+      output$table <- renderUI({
         .log("update uifiltersinfo")
         datanames <- if (identical(active_datanames(), "all")) {
           self$datanames()
@@ -726,8 +704,30 @@ FilteredData <- R6::R6Class( # nolint
           active_datanames()
         }
 
-        self$get_filter_overview_tbl(datanames = datanames)
-      }, width = "100%")
+        datasets_df <- self$get_filter_overview(datanames = datanames)
+
+        body_html <- lapply(seq_len(nrow(datasets_df)), function(x) {
+          tags$tr(
+            tags$td(rownames(datasets_df)[x]),
+            tags$td(datasets_df[x, 1]),
+            tags$td(datasets_df[x, 2])
+          )
+        })
+
+        header_html <- tags$tr(
+          tags$td(""),
+          tags$td(colnames(datasets_df)[1]),
+          tags$td(colnames(datasets_df)[2])
+        )
+
+        table_html <- tags$table(
+          class = "table custom-table",
+          tags$thead(header_html),
+          tags$tbody(body_html)
+        )
+
+        table_html
+      })
 
       return(invisible(NULL))
     }